Add files via upload
The analysis has been performed on a dataset which was merged from two distinct datasets, one of which contained Fake News along with the labels of them being fake and the other dataset contained True news along with the labels. The primary focus was to build a simple RNN that classifies fake and real news when randomly given either fake or real news.